Add type-level handling for flags.
This requires fleshing out the `type_set` class a bit. In addition, the
current design requires all `arg::value`s to be default-constructible.
For now I feel this isn't too onerous a restriction, but should probably
be revisited sometime in the future.

The shunting in of `short_name` to positional_arg seems wrong, too; this
should probably be removed or cleaned up in some way in a future commit.